Use LogEntry common entity
